Bank interest calculation
When a bank has financial problems, the government can return a client's deposit if it is less than 700,000. The interest rate for a particular deposit is 7.1% a year. Interest is added to the same deposit at the end of the year, and then a new value of the interest is calculated. Find out how many years it will take for the sum of the deposit to exceed the value protected by the government.